crudを分離
