--- .vscode/launch.json
@@ -1,28 +1,0 @@
-{
-    // IntelliSense を使用して利用可能な属性を学べます。
-    // 既存の属性の説明をホバーして表示します。
-    // 詳細情報は次を確認してください: https://go.microsoft.com/fwlink/?linkid=830387
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "name": "fueoni_ver2",
-            "cwd": "fueoni_ver2",
-            "request": "launch",
-            "type": "dart"
-        },
-        {
-            "name": "fueoni_ver2 (profile mode)",
-            "cwd": "fueoni_ver2",
-            "request": "launch",
-            "type": "dart",
-            "flutterMode": "profile"
-        },
-        {
-            "name": "fueoni_ver2 (release mode)",
-            "cwd": "fueoni_ver2",
-            "request": "launch",
-            "type": "dart",
-            "flutterMode": "release"
-        }
-    ]
-}
