+++ fueoni_backend/app/crud/__init__.py
@@ -1,0 +1,1 @@
+from app.crud.missions import *
