--- fueoni_backend/app/models/__init__.py
@@ -1,1 +1,0 @@
-from app.models.mission import *
